requirements.txt
streamlit==0.89.0
numpy==1.20.3
matplotlib==3.4.2
Pillow==8.2.0
